### 2017-04-25 #112 @crew102
patentsview
This package provides a wrapper around the [patentsview api](http://www.patentsview.org/api/doc.html), which serves USPTO patent data that has been disambiguated. There is one main function to search and download data from the API (`search_pv`) and several other helper functions.

```
Package: patentsview
Title: An R Client to the PatentsView API
Version: 0.1.0.9000
Authors@R: person("Christopher", "Baker", email = "chriscrewbaker@gmail.com",
    role = c("aut", "cre"))
Description: Provides functions to simplify the PatentsView API
    (http://www.patentsview.org/api/doc.html) query language,
    send GET and POST requests to the API's seven endpoints, and parse the data
    that comes back.
    httr,
    utils
```

[master branch on github](https://github.com/crew102/patentsview/tree/master)
Researchers who want to interact with patent data. There are several specific use cases of patent data that have been listed in [this thread](https://discuss.ropensci.org/t/ropensci-vs-ropengov/651/2) of the rOpenSci discussion board.

I know of two R packages that one can use to download patent data: [lensr](https://github.com/wipo-analytics/lensr) and a development version of [opsr](https://github.com/wipo-analytics/opsrdev). `lensr` is the most similar, as it can be used to download USPTO patent data. I do not know of any other packages that wrap patentsview, which is the only service that provides USPTO data that has been disambiguated.

`R CMD check` passes with one warning regarding this being my first package submission.

I believe the package follows the guidelines, but there are a few items I would like to highlight. These may or may not be issues:

1. The README does not follow the exact same format as described in the packaging guidelines.
2. I have added a cran-comment.md file which I plan on updating before submitting to CRAN.
3. I have not tagged the git repo yet, as I would like to wait for feedback before considering this version complete.
