### 2017-04-29 #112 @crew102
Thanks for running running those commands for me, @maelle. I have:
* Fixed spelling mistakes and added line breaks to code lines > 80 chars in length.
* Looked over all of my `sapply` calls and either changed to `vapply` or made sure that I had specified `simplify = FALSE`.
* Added additional unit tests.

The newest version of master is available here: crew102/patentsview@e6fdbc84fac6188e3ff1a2619c157199248e56eb
