### 2017-05-19 #112 @expectopatronum
Me too! So far I have one remark: I think that in the method `validate_endpoint` in utils.R the error message

> "endpoint must be a length 1 vector and be one of: "

might be confusing. I'm not sure if for an R novice it is clear that one element is equivalent to a vector of length one. Maybe the error message can changed to something like "endpoint must be a single element and be one of:"

I hope comments like this are ok here :)
