### 2017-05-23 #112 @expectopatronum
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

I'm very impressed by the code in the package, I think it is very well written and I enjoyed reviewing it! I only have a few remarks/comments.

* I installed and ran it successfully on Windows 10 (R-3.4.0) and Mac OS X (R-3.3.2)
* From my point of view the package follows most of the rOpenSci [packing recommendations](https://github.com/ropensci/packaging_guide) and follows all required things, in more detail:
  * It seems it does not generate the README.md from a README.rmd file and it does not include a code of conduct, however it is only a recommendation. I just wanted to point it out in case the author intends to change that.
  * It has the mandatory NEWS file.
  * Uses recommended packages (e.g. jsonlite)

**Possibilities for improvement**
* Testing is done for the main function `search_pv` and for error handling, which I think is very important. At first I thought there should be unit tests covering every exported function but after testing myself I think that it would be sufficient to add (a) test(s) for `flatten_pv_data()`.
* Confusing error message in `validate_endpoint` already mentioned in [my comment](https://github.com/ropensci/onboarding/issues/112#issuecomment-302781074)
* For convenience (since there is also `get_fields()`) a function `get_endpoints()` would be nice
* I tried to build my own queries and in one of them I tried to use `location_longitude` as a field for querying. I got an error **location_longitude is not a valid field to query
Error in if (f1$data_type == "date" && !is_date(value)) paste0_msg("Bad date: ",  :
  missing value where TRUE/FALSE needed** but I could not find out (using the help and the [vignette](https://github.com/crew102/patentsview/blob/master/vignettes/writing-queries.Rmd)) which fields are allowed for querying. Unless I missed something this should be added to the vignette/manual.
