### 2017-05-30 #112 @crew102
Hi @expectopatronum -

In addition to the notes/changes shown below, I have made the following three updates to `patentsview`:

* A new function called `get_ok_pk()` was added (https://github.com/crew102/patentsview/commit/0c4bbb91e24626d1ec82c71507066d9030dce3a4). `get_ok_pk()` suggests an appropriate primary key variable to use for the `pk_var` argument to `flatten_pv_data()`.

* The `pk_var` argument in `flatten_pv_data()` was renamed to `pk` (https://github.com/crew102/patentsview/commit/f363c7f83fb7ce3d67df6594d9a2aa6979655649).

* `flatten_pv_data()` was renamed to `unnest_pv_data()` (https://github.com/crew102/patentsview/commit/b9efa3d00b3e27522454f93b22bd9ec07a6e761e).

> It seems it does not generate the README.md from a README.rmd file

The markdown variant of README is actually generated from the rmarkdown variant, it just happens in a sort of atypical way. You can see the README.md target in the Makefile for details (https://github.com/crew102/patentsview/blob/master/Makefile#L8-L11).

> Testing is done for the main function search_pv and for error handling, which I think is very important. At first I thought there should be unit tests covering every exported function but after testing myself I think that it would be sufficient to add (a) test(s) for flatten_pv_data().

I have added a unit test for `unnest_pv_data()` (https://github.com/crew102/patentsview/commit/b9a20a1080c68106fad64b1588aef0b0c540a01c).

> Confusing error message in validate_endpoint already mentioned in my comment

I have changed the wording so it is not so confusing
(https://github.com/crew102/patentsview/commit/ed08c807e9459db98fdaa23e74cdbf9b2506c3dd).

> For convenience (since there is also get_fields()) a function get_endpoints() would be nice

Good idea...I have added `get_endpoints()` (https://github.com/crew102/patentsview/commit/ee3f65e9f32db288e913b6180473147aaa9699a8).

> I tried to build my own queries and in one of them I tried to use location_longitude as a field for querying. I got an error location_longitude is not a valid field to query Error in if (f1$data_type == "date" && !is_date(value)) paste0_msg("Bad date: ", : missing value where TRUE/FALSE needed but I could not find out (using the help and the vignette) which fields are allowed for querying. Unless I missed something this should be added to the vignette/manual.

There are some fields that you are not allowed to include in your query but can include in your `fields` argument (i.e., you can retrieve data about a field, but you can't filter on that field in your query). `get_fields()` returns all the fields that you can retrieve for a given endpoint, but some of those fields may not be valid to filter on in your query. It looks like `location_longitude` is actually not a queryable field, as per the [locations endpoint fields list](http://www.patentsview.org/api/location.html#field_list). Was the distinction between a queryable field and a retrievable field not clear from the documentation? If no, I will clarify.
