### 2017-06-01 #112 @crew102
> I tried to build my own queries and in one of them I tried to use location_longitude as a field for querying. I got an error location_longitude is not a valid field to query Error in if (f1$data_type == "date" && !is_date(value)) paste0_msg("Bad date: ", : missing value where TRUE/FALSE needed but I could not find out (using the help and the vignette) which fields are allowed for querying. Unless I missed something this should be added to the vignette/manual.

I have made the distinction between queryable vs retrievable fields clearer in the README (https://github.com/crew102/patentsview/commit/070a24186bebb3d0adc3a13ef38e4f775d6df313).
