### 2017-06-01 #112 @poldham
Dear @crew102,

Apologies for the delay... I was unexpectedly asked to do an overseas presentation and that took up a lot of time.

Before I get into the reviewing part, let me say that it is fantastic that this package has been prepared for R as it could be an incredible resource. Most of my comments will be on thinking about the user experience for those who are R users and who may or may not know much about patent data. The comments are all intended to be constructive. In a discussion last week with colleagues at WIPO we already plan to start using the package in the analytics trainings around the world this year. But, as you will see below, I have some comments on the user experience.

## Code Comments

My main code comment is on the requirement for users to write using `qry_funs$gt()`. The storage of the functions in a list as suggested by HW in Advanced R is really great but it does seem to get in the way of useability... since what the funs are is not clearly explained in the documentation. For example gt and gte must be blindingly obvious (get?) but I confess I remain mystified... and am fully expecting a duh moment.

One possiblility would be to try to mask the qry_funs$gt as simply `gt`. I found a discussion of options for making lists of functions less verbose at the end of this [section of Advanced R](http://adv-r.had.co.nz/Functional-programming.html#lists-of-functions).  It may be that @expectopatronum @maelle have some thoughts about this.

If it is not possible to find a way forward on that at the moment could I suggest that the list funs are clearly documented in `qry_funs()`.

### CMD Check

I receive a warning:

* checking package vignettes in ‘inst/doc’ ... WARNING
Package vignettes without corresponding PDF/HTML:
   ‘patentsview.Rmd’
   ‘writing-queries.Rmd’

The tests ran OK

### Vignettes

I think the package merits some work on the vignettes to make it easier for the first time user.

In both vignettes we see a description of moving from JSON to the DSL as the preferred approach. As the users will be coming in from R I think it makes sense to start with the DSL and maybe explain later on that there are other ways of doing this. I think (as an R user and sometime coder) that being confronted with

`search_pv(query = '{"_gt":{"patent_year":2010}}',
          method = "POST", fields = "patent_number",
          sort = c("patent_number" = "asc"))`

is a bit off putting  in term of engaging with the package and is actually, using the favoured DSL, not necessary.

There are of course a ton of things one could do with the package but could I suggest that the vignettes and examples focus on some of the main types of searches that patent researchers will normally do. Those would be:

1. Keyword and phrase searches
2. Year restricted searches (publication year and priority year)
3. Applicant (assignee) and inventor searches
4. IPC or CPC searches
5. Citations
6. Combinations of the above

In addition the locations endpoint is really really exciting (and I suspect folks will jump all over it with Leaflet and so on), but I found it difficult to get going with that.

In terms of the vignettes it is of course up to you. On the patent side of things my own efforts to be straightforward are illustrated for [lensr](https://github.com/poldham/lensr) and [opsr](https://github.com/poldham/opsrdev). Those were mainly written as part of the WIPO analytics training course for complete beginners. That may give you some ideas. In addition, rOpenSci packages such as [rplos](https://ropensci.org/tutorials/rplos_tutorial.html) have good walkthroughs.  @maelle monkeylearn walkthrough was really helpful for a newbiew as an other example. I think that makes all the difference in helping people to engage with the package.

### Getting to Grips with the Fields

I have seen the good addition of `get_endpoints()` but I also spent a lot of time trying to figure out the fields and there are a lot. I discovered the `fieldsdf.csv` in the data-raw folder and would suggest that you might put that as a `.rda` in data so that the user can call it to work out where they can go.

In the documentation for `get_fields()` I would suggest

1. Under endpoint. Bullet list the 7 endpoints to make clear what they are and cross link to `get_endpoints()`.
2. Under groups. Bullet the groups so they are easier to appreciate.
3. Note that the location group appears in `data-raw/fieldsdf`(see below) but was not listed in groups documentation or in the online [table](http://www.patentsview.org/api/patent.html#field_list)
4. For consistency, note that the online table and fieldsdf use the term group whereas the arg is groups. It is a trivial point but may be worth harmonising.
5. fieldsdf in data-raw does not contain the description field. Unless there is a good reason could this be added back in to that table to assist with understanding the fields. I can see that there is a case that it is hard to read a long text string in a df in R. But then users could also be given direct access to fieldsdf.

### Inventors and Assignees.

Can I suggest that an explanation is needed somewhere that these are cleaned up data fields. I am not sure if there is a published account of the clean up algorithm yet but I saw the video. Not vital right now but important to know that the data has been cleaned and that the raw is also available for purists.

Those are my main comments at the moment. I intend to take another look as I know there have already been some changes but wanted to be sure you received these given the unexpected delay.
