### 2017-06-01 #112 @maelle
@poldham many thanks for your review! How many hours did you approximately spend reviewing the package?

Even better to see you mention my *monkeylearn* package 😉 it benefited from a review in its time!

@crew102 perfect regarding the README then.
