### 2017-06-01 #112 @poldham
A pleasure @crew102 there is so much in this package that I think one of the challenges is working out strategies for getting started.

@maelle somewhere around 6 hours as I did a long stint, got dragged away, and then another stint today.
