### 2017-06-02 #112 @sckott
@poldham thx for your review - can you give an estimate of number of hrs spent reviewing?
