### 2017-06-05 #112 @crew102
Thanks again @poldham for reviewing the package. My responses to your comments are shown below.

### qry_funs

It sounds like you have two issues with the DSL - first, the fact that the functions are stored in a list and thus can be verbose to use, and second, the lack of documentation on the functions themselves.

1. **verbosity**. I think it makes sense to keep them in a list due to the possibility of name conflicts. However, I agree that requiring `qry_funs$` can get in the way of usability when your query is complex (for example, https://github.com/crew102/patentsview/blob/master/man/with_qfuns.Rd#L21-L29). I included the `with_qfuns` function in the hopes of reducing that problem. If you feel like this may not be sufficient for some users, I could add a note somewhere in the documentation letting users know that they can assign the functions into their global environment (or some other relevant environment) using `list2env(qry_funs, envir = globalenv())`.

2. **documentation**. I see how someone may not understand the meaning of the functions, as I did not explicitly provide definitions for all of them. I was hoping to just refer people to the patentsview API documentation in the "writing-queries' vignette for these details, but it looks like I didn't get the link exactly correct for this. I have corrected the link and added explicit documentation for each function in the `qry_funs()` documentation (crew102/patentsview@1b474665)

#### R CMD Check

> I receive a warning:
checking package vignettes in ‘inst/doc’ ... WARNING
Package vignettes without corresponding PDF/HTML:
‘patentsview.Rmd’
‘writing-queries.Rmd’

This should be fixed now.

#### Vignettes

> As the users will be coming in from R I think it makes sense to start with the DSL and maybe explain later on that there are other ways of doing this.

I agree that the DSL may be more natural for R users. However, I think the first example of `search_pv` in the README/"patentsview vignette" should use the JSON version of the query instead of the DSL version (i.e., `'{"_gte":{"patent_date":"2007-01-01"}}'` instead of `qry_funs$gte(patent_date = "2007-01-01")`). I think the DSL deserves immediate explanation, whereas a JSON string does not, and hence it would be less of an "information overload" if the first example used JSON.

> There are of course a ton of things one could do with the package but could I suggest that the vignettes and examples focus on some of the main types of searches that patent researchers will normally do. Those would be:

I have added a new example of searching for patents belonging to a given assignee (crew102/patentsview@8cf0c1dd). I think the REAMDE now has all of the example types that you mentioned, but let me know if you think a few more would be helpful.

### Getting to Grips with the Fields

> I have seen the good addition of get_endpoints() but I also spent alot of time trying to figure out the fields and there are a lot. I discovered the fieldsdf.csv in the data-raw folder and would suggest that you might put that as a .rda in data so that the user can call it to work out where they can go.

Similar to the case of the query functions, I was hoping to just refer the interested reader to the online field tables (e.g., in the README and `search_pv` docs). I did actually include the `fieldsdf` data frame as a data object for internal use (it was in R/sysdata.rda in the source code), so you could have accessed it with `patentsview:::fieldsdf`. The reason why I didn't export or document this was that I felt it would be redundant with the existing patentsview documentation. Anyway, I have:

* Exported `fieldsdf` so you should be able to access it with `patentsview::fieldsdf` or `data("fieldsdf")` (crew102/patentsview@7b86056f).
* Updated the README/patentsview vignette to include a section on fields. The availability of `fieldsdf` is noted in this section (crew102/patentsview@3bde26aa).

> In the documentation for get_fields() I would suggest
> Under endpoint. Bullet list the 7 endpoints to make clear what they are and cross link to get_endpoints().

I have linked to `get_endpoints()` in the `get_fields()` documentation (crew102/patentsview@c7647206). I think I would prefer to keep the endpoints and groups shown inline as they are now, as a list would be fairly long.

> Note that the location group appears in data-raw/fieldsdf(see below) but was not listed in groups documentation or in the online table

The location group does not exist in the patents endpoint, so that's probably why you're not seeing it online (assuming you are just looking at the patents table). It exists for locations, inventors, and assignees endpoints (e.g., http://www.patentsview.org/api/location.html#field_list)

>  For consistency, note that the online table and fieldsdf use the term group whereas the arg is groups. It is a trivial point but may be worth harmonising.

I typically like to use plurals for parameters which the user can specify multiple choices for. For example, `endpoint` is not plural b/c you can only specify one endpoint, but `fields` is plural b/c you can specify several fields.

> fieldsdf in data-raw does not contain the description field. Unless there is a good reason could this be added back in to that table to assist with understanding the fields. I can see that there is a case that it is hard to read a long text string in a df in R. But then users could also be given direct access to fieldsdf.

I have added "common name" and "description" to `fieldsdf`.

### Inventors and Assignees.

> Can I suggest that an explanation is needed somewhere that these are cleaned up data fields. I am not sure if there is a published account of the clean up algorithm yet but I saw the video. Not vital right now but important to know that the data has been cleaned and that the raw is also available for purists.

Yes, it is very important that people realize that they are not getting raw data by default. I have added additional explanation about this in the README/patentsview vignette (crew102/patentsview@938d3167).
