### 2017-06-08 #112 @expectopatronum
Yes absolutely, great work @crew102 and thanks for the great discussions. I really appreciate the clarification in the README and the additional test. @maelle Now I check the checkbox next to "The author has responded to my review and made changes to my satisfaction. I recommend approving this package." and that's it? Or are there any other steps?
