### 2017-06-09 #112 @crew102
Hi @maelle , I would like to submit `patentsview` to CRAN. Is the review process for ROpenSci over, or are we still waiting on @poldham to approve my changes?
