### 2017-06-10 #112 @poldham
@crew102 A somewhat mad week here in the UK and getting madder.

Many thanks for your responses on the review and here are my observations.

### qry_funs

1. On the qry_funs. I like the list approach with the only issue being verbosity. On with_qfuns, I think my issue was wouldn't it generally make more sense to use with_qfuns in the first place? I am not going to press on this but think the note on assigning could be helpful.

2. The update to the qry_funs documentation is exactly what was needed.

### CMD Check

1. The warning is fixed but a note is now appearing as:

Status: 1 NOTE
checking R code for possible problems ... NOTE
check_query: no visible binding for global variable ‘fieldsdf’
get_fields: no visible binding for global variable ‘fieldsdf’
validate_args: no visible binding for global variable ‘fieldsdf’
  fieldsdf

A bit puzzling because the documentation in data.R looks OK to me.

While on data.R could I suggest adding in the URL to the online version  of fieldsdf.

### Vignettes

1. As in my earlier comments I think starting with the JSON and moving the the DSL could be confusing and off-putting... notably for a new R user. As in, whoah... why am I being asked to input JSON strings into an R function... does not compute from the Coursera/Datacamp course I'm doing...get me out of here.  I mention this also because the survey of patent data analytics practices we ran for WIPO revealed that most people are still using Excel... sooo... However, a potential solution to this would be to add an explanatory comment at the beginning of the README under Basic Usage roughly to the effect that:

"The PatentsView API provides an interface to a disambiguated version of USPTO patent data. The PatentsView API uses JSON strings as described on the [PatentsView webpage](http://www.patentsview.org/api/patent.html). In the R package you can use JSON strings, as in the examples below, or you can use the simpler domain specific language (DSL) described below to make writing queries in R easier."

That would be my suggestion on this but I leave the precise wording to you. I do however think there needs to be a bit more explanatory lead in.

2. On the README examples, I think there are now a nice range of examples. The Fields discussion is good with the links to fieldsdf. The distinction between queryable and retrievable helps a lot.

3. I suggest another quick spell check on the README e.g. "aplied to"

4. On list columns in the results, that will be very helpful to users unfamiliar with list columns (and from my work on EPO data patent packages will probably always involve them). At the very end of the README on joining on the primary key you may wish to consider referencing see `dplyr::join()` for guidance on how to do this. Chapter 13 of HW & GGs R for Data Science [here](http://r4ds.had.co.nz/relational-data.html) also provides a very good practical walkthrough on joins in R.

5. On the link established between `get_fields()` and `get_endpoints()`. I can see the length issue on groups. I would propose to bullet the endpoints in get_fields for easy reference and leave the groups inline. I propose this because a user needs to understand what the endpoints are to be able to get their head around the groups.

6. On the location issue. Thanks and that is now clear to me. I will leave it with you on whether to document that clarification or not.

7. Ok on groups term in args and good explanation.

8. fieldsdf really is a great help.

9. On raw vs. disambiguated data this is now very clear. You might consider adding the reference to the disambiguation process e.g.  For those who might be interested in how the disambiguation was done by the team led by Nicholas Monath and Andrew McCallum from the University of Massachusetts Amherst the pdf is [here](http://www.patentsview.org/data/presentations/UMassInventorDisambiguation.pdf) and details of the 2015 PatentsView Inventor Disambiguation Technical Workshop on approaches to disambiguation are [here](http://www.patentsview.org/workshop/).

Subject to the above minor points @maelle @expectopatronum I am very happy.

@maelle please add another 1.5 hours to my review time.
