### 2017-06-12 #112 @crew102
Hi @expectopatronum , thanks for the feedback. Below you will find my responses. @maelle , where do we go from here? Should I get the package ready for submission to CRAN and just submit myself?

----------------------
> On with_qfuns, I think my issue was wouldn't it generally make more sense to use with_qfuns in the first place? I am not going to press on this but think the note on assigning could be helpful.

* You can use `with_qfuns` exclusively, but I think it makes sense to explain to users all of the possibilities and let them decide. I have added a note regarding how users can assign the functions into their global environment (crew102/patentsview@5e95770).

> The warning is fixed but a note is now appearing as:
Status: 1 NOTE
checking R code for possible problems ... NOTE
check_query: no visible binding for global variable ‘fieldsdf’
get_fields: no visible binding for global variable ‘fieldsdf’
validate_args: no visible binding for global variable ‘fieldsdf’
fieldsdf

* This should be fixed now (crew102/patentsview@e1a2f257).

> While on data.R could I suggest adding in the URL to the online version of fieldsdf.

* I have a URL to one of the endpoint's online field tables (crew102/patentsview@2e4ad0c)

> However, a potential solution to this would be to add an explanatory comment at the beginning of the README under Basic Usage roughly to the effect that:

* I have added a note to the README (crew102/patentsview@e25f222f05).

> I suggest another quick spell check on the README e.g. "aplied to"

* I have fixed this mistake and checked spelling again (crew102/patentsview@da220a5290).

> On list columns in the results, that will be very helpful to users unfamiliar with list columns (and from my work on EPO data patent packages will probably always involve them). At the very end of the README on joining on the primary key you may wish to consider referencing see dplyr::join() for guidance on how to do this. Chapter 13 of HW & GGs R for Data Science here also provides a very good practical walkthrough on joins in R.

I have referenced the Relational data chapter in the documentation for `unnest_pv_data` (crew102/patentsview@fa9a406df).

> On the link established between get_fields() and get_endpoints(). I can see the length issue on groups. I would propose to bullet the endpoints in get_fields for easy reference and leave the groups inline. I propose this because a user needs to understand what the endpoints are to be able to get their head around the groups.

* I have added a bulleted list with the endpoints to the `get_endpoints()` documentation (crew102/patentsview@e4a140ac6). I think the endpoints should be listed here instead of at `get_fields()`. A link to `get_endpoints()` is given in the documentation for `get_fields()`.

> You might consider adding the reference to the disambiguation process

* I think that referencing the PatentsView online documentation should suffice for this.
