### 2017-06-12 #112 @maelle
@poldham are you happy with this response?

@crew102 I'll do the last checks later this week, thanks for your patience! 😊 After the last checks You'll transfer your repo to ropensci therefore the URLs in DESCRIPTION  will change so please wait before the CRAN submission. 😉
