### 2017-06-12 #112 @poldham
@crew102 @maelle.  Thanks @crew102. This is all fine with me except that I still think that the opening sentence of the README as raised under point 1 in Vignettes above still needs a sentence or so to better orient the reader that a) they are dealing with a JSON API and b) that there is an easy to use DSL.

On the disambiguation point I am assuming that you mean [http://www.patentsview.org/workshop/](http://www.patentsview.org/workshop/) which is fine with me. I think the reference to the disambiguation method process is important to give users confidence in using the disambiguated data as inventor disambiguation has been an intractable and non-trivial problem for many many years.

Looking forward to seeing this join the ropensci collection and to spreading the word and using it. Many congratulations!
