### 2017-06-15 #112 @maelle
@crew102 approved! :smile_cat:

I've added you to the ropensci organization, could you transfer the repository? After that you'll have to change all URLs and I'll activate continuous integration via rOpenSci account (if not today, then tomorrow).

Besides, please add the rOpenSci footer `[![ropensci_footer](http://ropensci.org/public_images/github_footer.png)](http://ropensci.org)` to the bottom of the README.
