### 2017-06-16 #112 @crew102
`rmarkdown::render` was having a hard time finding http://ropensci.org/public_images/github_footer.png (not sure why, though...Any idea what ), which was causing my makefile to fail when I added the footer to patentsview.rmd (which is the readme). I figured it'd just append the raw readme.md file with the footer text like so:
https://github.com/crew102/patentsview/blob/master/Makefile#L12
It works but you have to run the makefile in order to refresh the readme.me. Is that going to be a problem for ropensci? Running the readme will fail if you don't have Rscript on your path.
