### 2017-06-16 #112 @crew102
Also, does ropensci have an AppVeyor account? If so, can we run builds on AppVeyor as well as Travis?
