### 2017-06-16 #112 @maelle
Oops sorry yes this is a current pandoc bug so you can wait until the pandoc bug is fixed to add itn (will tell you)

Yes I'll activate your repo on appveyor as well😊
