### 2017-06-16 #112 @crew102
Can you confirm that I have permission to create repos on ropensci? Github thinks I don't have permission.
