### 2017-06-19 #112 @maelle
Have you received my invitation to the patentsview team in the rOpenSci organization @crew102 ?
