### 2017-06-19 #112 @crew102
Hi @maelle , I haven't seen any invitations go by in my notifications. I get this message when I try to transfer the repo to ropensci:

![x](https://user-images.githubusercontent.com/11220793/27290244-a1503f52-54da-11e7-908a-8cd687f4b347.png)
