### 2017-06-19 #112 @maelle
@crew102 just check the patentsview team.

Thanks @sckott for guidance (@crew102 disclaimer: your package is the first one I've ever accepted, so thanks for your patience)
