### 2017-06-19 #112 @crew102
No problem - I thought the process was well done. I'm changing the urls now and will open a pr to integrate those into ropensci/patentsview. I'm also changing the badge urls for travis and appveyor.
