### 2017-06-19 #112 @maelle
Cool!

The repo was activated on Travis and I think on Appveyor for which the badges will be `[![Build status](https://ci.appveyor.com/api/projects/status/ri4lrakyrw2117mi?svg=true)](https://ci.appveyor.com/project/ropensci/patentsview)`

No need for a PR, you can commit and push directly well unless you prefer doing it with a branch.
