### 2017-06-19 #112 @crew102
Sounds good..I did a PR but I'll probably just push directly to ropensci/patentsview master in the future. Thanks again for your help.
