### 2017-06-20 #112 @crew102
Hi @maelle , is it ok if I submit patentsview to CRAN without the ropensci footer in the readme? I'm getting this warning when I build on winbuilder (which I assume will be a showstopper for CRAN admins):
````
checking top-level files ... WARNING
Conversion of 'README.md' failed:
pandoc.exe: Could not fetch http://ropensci.org/public_images/github_footer.png
TlsExceptionHostPort (HandshakeFailed Error_EOF) "ropensci.org" 80
````
