### 2017-06-20 #112 @maelle
@crew102 yes with that Pandoc bug the best is to remove it before CRAN submission and then put it back afterwards. Hopefully that bug will be resolved soon.

Would you be interested in writing a blog post about `patentsview` for rOpenSci blog?
