### 2017-06-20 #112 @crew102
Sure, I could do a post. I'd like to submit to CRAN first. I'll let you know when that process is over. Are there any guidelines/suggestions for writing the blog posts?
