### 2017-06-20 #112 @maelle
Awesome! I'll let @stefaniebutland take over reg. the blog post (she's rOpenSci community manager), she'll be able to tell you more regarding timing & content of the blog post.

In any case I look forward to reading it!

Good luck with the CRAN submission, don't hesitate to ask any question if needed.
